move vect tests
